Create run_experiment.py
Set up the run_experiment file. I intend for this to be the "admin" executable. It should be called or run given some configuration. 

I'm going to build out the individual components here - it may make sense to break them out as separate function but I think this is the most straightforward place to start.